+++ examples/apps/lit-router/CHANGELOG.md
@@ -1,5 +1,11 @@
 # lit-todomvc
 
+## 0.2.2
+
+### Patch Changes
+
+- @inglorious/lit@1.2.1
+
 ## 0.2.1
 
 ### Patch Changes