+++ examples/apps/react-todomvc-ts/CHANGELOG.md
@@ -1,5 +1,12 @@
 # todomvc
 
+## 0.1.7
+
+### Patch Changes
+
+- @inglorious/store@7.0.0
+- @inglorious/react-store@8.0.0
+
 ## 0.1.6
 
 ### Patch Changes