+++ packages/server/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @inglorious/server
 
+## 10.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/utils@3.7.0
+  - @inglorious/store@7.0.0
+
 ## 9.0.1
 
 ### Patch Changes