+++ packages/store/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @inglorious/store
 
+## 7.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/utils@3.7.0
+
 ## 6.2.1
 
 ### Patch Changes